+++ PiesGenericPatch/ModFolderContent/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.2.0
+
+- Fix citizen needs (bladder especially) increasing many times faster than intended while they interact or talk to the player
+
+# 1.1.0
+
+- Prevent loaded cities from spawning credit cards and blood donor cards in random folders
+
 # 1.0.0
- 
-* Initial Version 
+
+- Initial Version